refactor(cli)!: retire repo commands
Route the legacy namespace to project-mode migration guidance and keep\ndaemon upgrades safe by respawning and stopping through standalone project\nidentity.

Closes #849
