+++ .changeset/retire-repo-command.md
@@ -1,0 +1,5 @@
+---
+"@gh-symphony/cli": major
+---
+
+Retire the `repo` command surface in favor of explicit `project start --project-dir <path>` orchestration, including safe daemon respawn and legacy daemon stopping (#849).
