--- packages/cli/src/commands/repo.test.ts
@@ -1,1005 +1,0 @@
-import { execFileSync } from "node:child_process";
-import {
-  chmod,
-  mkdir,
-  mkdtemp,
-  readFile,
-  stat,
-  writeFile,
-} from "node:fs/promises";
-import { join } from "node:path";
-import { tmpdir } from "node:os";
-import { afterEach, describe, expect, it, vi } from "vitest";
-import type { CliProjectConfig } from "../config.js";
-import {
-  loadGlobalConfig,
-  loadProjectConfig,
-  saveGlobalConfig,
-} from "../config.js";
-import {
-  inspectManagedProjectSelection,
-  resolveManagedProjectConfig,
-} from "../project-selection.js";
-import { standaloneProjectId } from "../standalone-project.js";
-
-async function loadRepoCommand() {
-  vi.resetModules();
-  const mod = await import("./repo.js");
-  return mod.default;
-}
-
-async function loadRepoModule() {
-  vi.resetModules();
-  return import("./repo.js");
-}
-
-async function loadRepoRuntimeModule() {
-  vi.resetModules();
-  return import("../repo-runtime.js");
-}
-
-function captureWrites(stream: NodeJS.WriteStream): {
-  output: () => string;
-  restore: () => void;
-} {
-  let buffer = "";
-  const spy = vi.spyOn(stream, "write").mockImplementation(((
-    chunk: string | Uint8Array
-  ) => {
-    buffer +=
-      typeof chunk === "string" ? chunk : Buffer.from(chunk).toString("utf8");
-    return true;
-  }) as typeof stream.write);
-
-  return {
-    output: () => buffer,
-    restore: () => spy.mockRestore(),
-  };
-}
-
-function baseOptions(configDir: string) {
-  return {
-    configDir,
-    verbose: false,
-    json: false,
-    noColor: true,
-  };
-}
-
-const VALID_WORKFLOW = `---
-tracker:
-  kind: github-project
-  provider:
-    project_id: PVT_project_123
-    state_field: Status
-  active_states:
-    - Ready
-  terminal_states:
-    - Done
-polling:
-  interval_ms: 30000
-workspace:
-  root: .runtime/symphony-workspaces
-hooks:
-  timeout_ms: 60000
-agent:
-  max_concurrent_agents: 1
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`;
-
-const FILE_WORKFLOW = `---
-tracker:
-  kind: file
-  provider:
-    path: $GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH
-    project_id: e2e-test
-codex:
-  command: codex app-server
----
-Prompt body.
-`;
-
-async function createGitRepo(remoteName = "platform"): Promise<string> {
-  const repoDir = await mkdtemp(join(tmpdir(), "repo-init-"));
-  execFileSync("git", ["-C", repoDir, "init"]);
-  execFileSync("git", [
-    "-C",
-    repoDir,
-    "remote",
-    "add",
-    "origin",
-    `https://github.com/acme/${remoteName}.git`,
-  ]);
-  return repoDir;
-}
-
-async function readRepoProjectConfig(
-  repoDir: string
-): Promise<CliProjectConfig> {
-  return JSON.parse(
-    await readFile(
-      join(
-        repoDir,
-        ".runtime",
-        "orchestrator",
-        "projects",
-        "repository",
-        "project.json"
-      ),
-      "utf8"
-    )
-  ) as CliProjectConfig;
-}
-
-afterEach(() => {
-  vi.restoreAllMocks();
-  process.exitCode = undefined;
-});
-
-describe("repo init runtime migration", () => {
-  it("writes repository selection to an explicit config directory", async () => {
-    const repoDir = await createGitRepo();
-    const configDir = await mkdtemp(join(tmpdir(), "repo-init-config-dir-"));
-    const repoCommand = await loadRepoCommand();
-    await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-    await saveGlobalConfig(configDir, {
-      activeProject: "standalone",
-      projects: ["standalone"],
-    });
-
-    await repoCommand(["init", "--repo-dir", repoDir], {
-      ...baseOptions(configDir),
-      configDirOverride: true,
-    });
-
-    const projectId = `repo-${standaloneProjectId(repoDir)}`;
-    await expect(loadGlobalConfig(configDir)).resolves.toEqual({
-      activeProject: projectId,
-      projects: ["standalone", projectId],
-    });
-    await expect(
-      loadProjectConfig(configDir, projectId)
-    ).resolves.toMatchObject({
-      projectId,
-      repository: { owner: "acme", name: "platform" },
-    });
-    await expect(
-      inspectManagedProjectSelection({ configDir, cwd: repoDir })
-    ).resolves.toMatchObject({
-      kind: "resolved",
-      projectId,
-      projectConfig: {
-        repository: { owner: "acme", name: "platform" },
-      },
-    });
-    await expect(
-      resolveManagedProjectConfig({ configDir, cwd: repoDir })
-    ).resolves.toMatchObject({
-      projectId,
-      repository: { owner: "acme", name: "platform" },
-    });
-  });
-
-  it("keeps repository records distinct in a shared explicit config directory", async () => {
-    const repoA = await createGitRepo("platform-a");
-    const repoB = await createGitRepo("platform-b");
-    const configDir = await mkdtemp(join(tmpdir(), "repo-init-shared-config-"));
-    const { initRepoRuntime } = await loadRepoRuntimeModule();
-    await Promise.all(
-      [repoA, repoB].map((repoDir) =>
-        writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8")
-      )
-    );
-
-    await initRepoRuntime({ repoDir: repoA, configDir });
-    await initRepoRuntime({ repoDir: repoB, configDir });
-
-    await expect(
-      inspectManagedProjectSelection({ configDir, cwd: repoA })
-    ).resolves.toMatchObject({
-      kind: "resolved",
-      projectId: `repo-${standaloneProjectId(repoA)}`,
-      projectConfig: {
-        repository: { owner: "acme", name: "platform-a" },
-      },
-    });
-    await expect(
-      loadProjectConfig(configDir, `repo-${standaloneProjectId(repoA)}`)
-    ).resolves.toMatchObject({
-      repository: { owner: "acme", name: "platform-a" },
-    });
-    await expect(
-      inspectManagedProjectSelection({ configDir, cwd: repoB })
-    ).resolves.toMatchObject({
-      kind: "resolved",
-      projectId: `repo-${standaloneProjectId(repoB)}`,
-      projectConfig: {
-        repository: { owner: "acme", name: "platform-b" },
-      },
-    });
-    await expect(loadGlobalConfig(configDir)).resolves.toMatchObject({
-      projects: [
-        `repo-${standaloneProjectId(repoA)}`,
-        `repo-${standaloneProjectId(repoB)}`,
-      ],
-    });
-  });
-
-  it("preserves an explicit global config override for repo subcommands", async () => {
-    const { repoOptions } = await loadRepoModule();
-    const configDir = join(tmpdir(), "captured-runtime");
-
-    expect(
-      repoOptions({
-        ...baseOptions(configDir),
-        configDirOverride: true,
-      }).configDir
-    ).toBe(configDir);
-  });
-
-  it("preserves a non-default environment config override for repo subcommands", async () => {
-    const { repoOptions } = await loadRepoModule();
-    const configDir = join(tmpdir(), "captured-runtime-env");
-
-    expect(
-      repoOptions({
-        ...baseOptions(configDir),
-        configDirOverride: true,
-        configDirSource: "env",
-      }).configDir
-    ).toBe(configDir);
-  });
-
-  it("keeps the cwd repo runtime for the official Docker default env path", async () => {
-    const { repoOptions } = await loadRepoModule();
-    const repoDir = await mkdtemp(join(tmpdir(), "repo-options-docker-"));
-    const repoRuntimeDir = join(repoDir, ".runtime", "orchestrator");
-    const originalCwd = process.cwd();
-
-    await mkdir(repoRuntimeDir, { recursive: true });
-    await writeFile(
-      join(repoRuntimeDir, "config.json"),
-      JSON.stringify({ activeProject: "repository", projects: ["repository"] }),
-      "utf8"
-    );
-
-    try {
-      process.chdir(repoDir);
-      const expectedRepoRuntimeDir = join(
-        process.cwd(),
-        ".runtime",
-        "orchestrator"
-      );
-      expect(
-        repoOptions({
-          ...baseOptions("/var/lib/gh-symphony"),
-          configDirOverride: true,
-          configDirSource: "env",
-        }).configDir
-      ).toBe(expectedRepoRuntimeDir);
-    } finally {
-      process.chdir(originalCwd);
-    }
-  });
-
-  it("falls back to the cwd repo runtime when no config override was supplied", async () => {
-    const { repoOptions } = await loadRepoModule();
-    const repoDir = await mkdtemp(join(tmpdir(), "repo-options-"));
-    const originalCwd = process.cwd();
-
-    try {
-      process.chdir(repoDir);
-      expect(repoOptions(baseOptions(join(tmpdir(), "global"))).configDir).toBe(
-        join(process.cwd(), ".runtime", "orchestrator")
-      );
-    } finally {
-      process.chdir(originalCwd);
-    }
-  });
-
-  it("does not advertise removed repo-set commands in fallback usage", async () => {
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-
-    try {
-      await repoCommand(["unknown"], baseOptions(tmpdir()));
-    } finally {
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(2);
-    expect(stderr.output()).toContain(
-      "Usage: gh-symphony repo <init|start|status|stop|run|recover|logs|explain> [repo]"
-    );
-    expect(stderr.output()).not.toContain("list|add|remove|sync");
-  });
-
-  it("promotes a single legacy project directory and strips projectId from run records", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-
-    const runDir = join(
-      repoDir,
-      ".runtime",
-      "orchestrator",
-      "projects",
-      "tenant-a",
-      "runs",
-      "run-1"
-    );
-    await mkdir(runDir, { recursive: true });
-    await writeFile(
-      join(runDir, "run.json"),
-      JSON.stringify({
-        runId: "run-1",
-        projectId: "tenant-a",
-        status: "running",
-      }),
-      "utf8"
-    );
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stdout.restore();
-    }
-
-    const migratedRun = JSON.parse(
-      await readFile(
-        join(repoDir, ".runtime", "orchestrator", "runs", "run-1", "run.json"),
-        "utf8"
-      )
-    ) as Record<string, unknown>;
-    const projectConfig = await readRepoProjectConfig(repoDir);
-
-    expect(migratedRun).not.toHaveProperty("projectId");
-    expect(projectConfig.repository).toMatchObject({
-      owner: "acme",
-      name: "platform",
-    });
-    expect(projectConfig.repositoryDir).toBe(repoDir);
-    expect(projectConfig.workspaceDir).toBe(
-      join(repoDir, ".runtime", "symphony-workspaces")
-    );
-    expect((await stat(projectConfig.workspaceDir)).isDirectory()).toBe(true);
-    expect((await stat(projectConfig.workspaceDir)).mode & 0o777).toBe(0o700);
-    expect(projectConfig).not.toHaveProperty("repositories");
-    expect(stdout.output()).toContain("Repository initialized: acme/platform");
-  });
-
-  it("can run repo init again after the repository layout exists", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-      process.exitCode = undefined;
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stdout.restore();
-    }
-
-    expect(process.exitCode).toBeUndefined();
-    expect(stdout.output()).toContain("Repository initialized: acme/platform");
-  });
-
-  it("uses a dedicated workspace root when workspace.root is omitted", async () => {
-    const repoDir = await createGitRepo();
-    const { initRepoRuntime } = await loadRepoRuntimeModule();
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      VALID_WORKFLOW.replace(
-        "workspace:\n  root: .runtime/symphony-workspaces\n",
-        ""
-      ),
-      "utf8"
-    );
-
-    await initRepoRuntime({ repoDir });
-
-    const projectConfig = await readRepoProjectConfig(repoDir);
-    expect(projectConfig.workspaceDir).toBe(
-      join(repoDir, ".runtime", "symphony-workspaces")
-    );
-    expect(projectConfig.workspaceDir).not.toBe(
-      join(repoDir, ".runtime", "orchestrator")
-    );
-  });
-
-  it("persists the explicit repo workflow file for daemon loading", async () => {
-    const repoDir = await createGitRepo();
-    const workflowPath = join(repoDir, "policy", "custom-workflow.md");
-    const { initRepoRuntime } = await loadRepoRuntimeModule();
-    await mkdir(join(repoDir, "policy"), { recursive: true });
-    await writeFile(workflowPath, VALID_WORKFLOW, "utf8");
-
-    await initRepoRuntime({
-      repoDir,
-      workflowFile: "policy/custom-workflow.md",
-    });
-
-    expect((await readRepoProjectConfig(repoDir)).workflowSource).toEqual({
-      type: "repo",
-      path: workflowPath,
-    });
-  });
-
-  it("enforces private permissions on a pre-existing workspace root", async () => {
-    const repoDir = await createGitRepo();
-    const workspaceDir = join(repoDir, ".runtime", "symphony-workspaces");
-    const { initRepoRuntime } = await loadRepoRuntimeModule();
-    await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-    await mkdir(workspaceDir, { recursive: true, mode: 0o755 });
-    await chmod(workspaceDir, 0o755);
-
-    await initRepoRuntime({ repoDir });
-
-    expect((await stat(workspaceDir)).mode & 0o777).toBe(0o700);
-  });
-
-  it("rejects workspace.root when it contains orchestrator state", async () => {
-    const repoDir = await createGitRepo();
-    const { initRepoRuntime } = await loadRepoRuntimeModule();
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      VALID_WORKFLOW.replace(".runtime/symphony-workspaces", ".runtime"),
-      "utf8"
-    );
-
-    await expect(initRepoRuntime({ repoDir })).rejects.toThrow(
-      "must not equal or contain the orchestrator runtime directory"
-    );
-  });
-
-  it.each([".", ".."])(
-    "rejects workspace.root %s when it equals or contains the checkout",
-    async (workspaceRoot) => {
-      const repoDir = await createGitRepo();
-      const { initRepoRuntime } = await loadRepoRuntimeModule();
-      await writeFile(
-        join(repoDir, "WORKFLOW.md"),
-        VALID_WORKFLOW.replace(".runtime/symphony-workspaces", workspaceRoot),
-        "utf8"
-      );
-
-      await expect(initRepoRuntime({ repoDir })).rejects.toThrow(
-        "must not equal or contain the repository checkout"
-      );
-    }
-  );
-
-  it("prints a friendly remediation when repo init cannot find WORKFLOW.md", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain("WORKFLOW.md was not found");
-    expect(stderr.output()).toContain("gh-symphony workflow init");
-    expect(stderr.output()).not.toContain("ENOENT");
-  });
-
-  it("prints JSON errors for repo init failures when --json is set", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-
-    try {
-      await repoCommand(["init", "--repo-dir", repoDir], {
-        ...baseOptions(join(repoDir, "unused")),
-        json: true,
-      });
-    } finally {
-      stdout.restore();
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toBe("");
-    expect(JSON.parse(stdout.output())).toEqual({
-      error: {
-        code: "missing_workflow_file",
-        message: expect.stringContaining("gh-symphony workflow init"),
-      },
-    });
-  });
-
-  it("persists explicit priority mapping from WORKFLOW.md into tracker settings", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      `---
-tracker:
-  kind: github-project
-  provider:
-    project_id: PVT_project_123
-    state_field: Status
-    priority_field: Legacy Priority
-    priority:
-      source: labels
-      labels:
-        P0: 0
-        P1: 1
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`,
-      "utf8"
-    );
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stdout.restore();
-    }
-
-    const projectConfig = await readRepoProjectConfig(repoDir);
-    expect(projectConfig.tracker.priority).toEqual({
-      source: "labels",
-      labels: {
-        P0: 0,
-        P1: 1,
-      },
-    });
-    expect(projectConfig.tracker.settings?.priorityFieldName).toBe(
-      "Legacy Priority"
-    );
-  });
-
-  it("persists a clean GitHub project id when WORKFLOW.md uses an inline comment", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      VALID_WORKFLOW.replace(
-        "project_id: PVT_project_123",
-        "project_id: PVT_kwHOAPiKdM4BYPVD # Moncher Stack (hojinzs/projects/14)"
-      ),
-      "utf8"
-    );
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stdout.restore();
-    }
-
-    const projectConfig = await readRepoProjectConfig(repoDir);
-
-    expect(projectConfig.tracker.bindingId).toBe("PVT_kwHOAPiKdM4BYPVD");
-    expect(projectConfig.tracker.settings?.projectId).toBe(
-      "PVT_kwHOAPiKdM4BYPVD"
-    );
-  });
-
-  it("initializes a Linear tracker runtime from WORKFLOW.md", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    const originalLinearApiKey = process.env.LINEAR_API_KEY;
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      `---
-tracker:
-  kind: linear
-  provider:
-    api_key: $LINEAR_API_KEY
-    project_slug: symphony-0c79b11b75ea
-    pickup_labels:
-      include:
-        - agent
-        - dev-ready
-      exclude:
-        - no-agent
-  active_states:
-    - Todo
-    - In Progress
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`,
-      "utf8"
-    );
-    process.env.LINEAR_API_KEY = "lin_test_token";
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      if (originalLinearApiKey === undefined) {
-        delete process.env.LINEAR_API_KEY;
-      } else {
-        process.env.LINEAR_API_KEY = originalLinearApiKey;
-      }
-      stdout.restore();
-    }
-
-    const projectConfig = await readRepoProjectConfig(repoDir);
-
-    expect(process.exitCode).toBeUndefined();
-    expect(projectConfig.tracker).toMatchObject({
-      adapter: "linear",
-      bindingId: "symphony-0c79b11b75ea",
-      apiUrl: "https://api.linear.app/graphql",
-      settings: {
-        projectSlug: "symphony-0c79b11b75ea",
-        activeStates: "Todo\nIn Progress",
-        pickupLabels: {
-          include: ["agent", "dev-ready"],
-          exclude: ["no-agent"],
-        },
-        repository: "acme/platform",
-      },
-    });
-  });
-
-  it("fails Linear repo init when tracker.api_key is missing", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      `---
-tracker:
-  kind: linear
-  provider:
-    project_slug: symphony-0c79b11b75ea
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`,
-      "utf8"
-    );
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain(
-      'Linear tracker repo init requires WORKFLOW.md field "tracker.api_key"'
-    );
-  });
-
-  it("rejects a missing Linear project_slug through adapter validation", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-    const originalLinearApiKey = process.env.LINEAR_API_KEY;
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      `---
-tracker:
-  kind: linear
-  provider:
-    api_key: $LINEAR_API_KEY
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`,
-      "utf8"
-    );
-    process.env.LINEAR_API_KEY = "lin_test_token";
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      if (originalLinearApiKey === undefined) {
-        delete process.env.LINEAR_API_KEY;
-      } else {
-        process.env.LINEAR_API_KEY = originalLinearApiKey;
-      }
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain(
-      "project_slug is required by the Linear tracker adapter."
-    );
-  });
-
-  it("fails Linear repo init when LINEAR_API_KEY reference is unresolved", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-    const originalLinearApiKey = process.env.LINEAR_API_KEY;
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      `---
-tracker:
-  kind: linear
-  provider:
-    api_key: $LINEAR_API_KEY
-    project_slug: symphony-0c79b11b75ea
-codex:
-  command: codex app-server
----
-Handle {{issue.identifier}}.
-`,
-      "utf8"
-    );
-    delete process.env.LINEAR_API_KEY;
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      if (originalLinearApiKey === undefined) {
-        delete process.env.LINEAR_API_KEY;
-      } else {
-        process.env.LINEAR_API_KEY = originalLinearApiKey;
-      }
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain(
-      'Workflow front matter field "tracker.provider.api_key" requires environment variable LINEAR_API_KEY'
-    );
-  });
-
-  it("resolves a relative --workflow-file against --repo-dir", async () => {
-    const repoDir = await createGitRepo();
-    const elsewhere = await mkdtemp(join(tmpdir(), "repo-init-elsewhere-"));
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    const originalCwd = process.cwd();
-    await writeFile(join(repoDir, "CUSTOM.md"), VALID_WORKFLOW, "utf8");
-
-    try {
-      process.chdir(elsewhere);
-      await repoCommand(
-        ["init", "--repo-dir", repoDir, "--workflow-file", "CUSTOM.md"],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      process.chdir(originalCwd);
-      stdout.restore();
-    }
-
-    expect(process.exitCode).toBeUndefined();
-    expect(stdout.output()).toContain(
-      `Workflow: ${join(repoDir, "CUSTOM.md")}`
-    );
-  });
-
-  it.each([
-    ["name.with.dots", "name.with.dots"],
-    [".github", ".github"],
-  ])(
-    "infers GitHub repository names containing dots: %s",
-    async (remoteName, expectedName) => {
-      const repoDir = await createGitRepo(remoteName);
-      const stdout = captureWrites(process.stdout);
-      const repoCommand = await loadRepoCommand();
-      await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-
-      try {
-        await repoCommand(
-          ["init", "--repo-dir", repoDir],
-          baseOptions(join(repoDir, "unused"))
-        );
-      } finally {
-        stdout.restore();
-      }
-
-      const projectConfig = await readRepoProjectConfig(repoDir);
-
-      expect(projectConfig.repository).toMatchObject({
-        owner: "acme",
-        name: expectedName,
-      });
-      expect(stdout.output()).toContain(
-        `Repository initialized: acme/${expectedName}`
-      );
-    }
-  );
-
-  it("fails with manual cleanup guidance when multiple legacy project directories exist", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-    await writeFile(join(repoDir, "WORKFLOW.md"), VALID_WORKFLOW, "utf8");
-    await mkdir(
-      join(repoDir, ".runtime", "orchestrator", "projects", "tenant-a"),
-      { recursive: true }
-    );
-    await mkdir(
-      join(repoDir, ".runtime", "orchestrator", "projects", "tenant-b"),
-      { recursive: true }
-    );
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain(
-      "Multiple legacy project runtime directories"
-    );
-    expect(stderr.output()).toContain("Manually keep the project directory");
-  });
-
-  it("reports a clear error when --project-id is used", async () => {
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-
-    try {
-      await repoCommand(
-        ["init", "--project-id", "tenant-a"],
-        baseOptions(tmpdir())
-      );
-    } finally {
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(2);
-    expect(stderr.output()).toContain("--project-id has been removed");
-    expect(stderr.output()).toContain("current repository directory");
-  });
-
-  it("writes a single-repository file-tracker config for repo-local E2E init", async () => {
-    const repoDir = await createGitRepo();
-    const stdout = captureWrites(process.stdout);
-    const repoCommand = await loadRepoCommand();
-    const originalIssuesPath = process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-    await writeFile(join(repoDir, "WORKFLOW.md"), FILE_WORKFLOW, "utf8");
-    process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH =
-      "/e2e/fixtures/issues.json";
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      if (originalIssuesPath === undefined) {
-        delete process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-      } else {
-        process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH = originalIssuesPath;
-      }
-      stdout.restore();
-    }
-
-    const projectConfig = await readRepoProjectConfig(repoDir);
-
-    expect(projectConfig.repository).toMatchObject({
-      owner: "acme",
-      name: "platform",
-    });
-    expect(projectConfig).not.toHaveProperty("repositories");
-    expect(projectConfig.tracker).toMatchObject({
-      adapter: "file",
-      bindingId: "e2e-test",
-      settings: {
-        projectId: "e2e-test",
-        repository: "acme/platform",
-        issuesPath: "/e2e/fixtures/issues.json",
-      },
-    });
-    expect(stdout.output()).toContain("Repository initialized: acme/platform");
-  });
-
-  it("fails fast when file-tracker repo init has no issues fixture path", async () => {
-    const repoDir = await createGitRepo();
-    const stderr = captureWrites(process.stderr);
-    const repoCommand = await loadRepoCommand();
-    const originalIssuesPath = process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      FILE_WORKFLOW.replace(
-        "    path: $GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH\n    project_id: e2e-test\n",
-        ""
-      ),
-      "utf8"
-    );
-    delete process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-    } finally {
-      if (originalIssuesPath === undefined) {
-        delete process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-      } else {
-        process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH = originalIssuesPath;
-      }
-      stderr.restore();
-    }
-
-    expect(process.exitCode).toBe(1);
-    expect(stderr.output()).toContain(
-      'File tracker requires "tracker.provider.path" or GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH.'
-    );
-  });
-
-  it("uses the legacy file fixture environment fallback when provider.path is absent", async () => {
-    const repoDir = await createGitRepo();
-    const repoCommand = await loadRepoCommand();
-    const originalIssuesPath = process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-    await writeFile(
-      join(repoDir, "WORKFLOW.md"),
-      FILE_WORKFLOW.replace(
-        "    path: $GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH\n    project_id: e2e-test\n",
-        ""
-      ),
-      "utf8"
-    );
-    process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH =
-      "/tmp/legacy-issues.json";
-
-    try {
-      await repoCommand(
-        ["init", "--repo-dir", repoDir],
-        baseOptions(join(repoDir, "unused"))
-      );
-      await expect(readRepoProjectConfig(repoDir)).resolves.toMatchObject({
-        tracker: { settings: { issuesPath: "/tmp/legacy-issues.json" } },
-      });
-    } finally {
-      if (originalIssuesPath === undefined) {
-        delete process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH;
-      } else {
-        process.env.GH_SYMPHONY_FILE_TRACKER_ISSUES_PATH = originalIssuesPath;
-      }
-    }
-  });
-});
