--- packages/cli/src/commands/repo.ts
@@ -1,157 +1,0 @@
-import type { GlobalOptions } from "../index.js";
-import { existsSync } from "node:fs";
-import logsCommand from "./logs.js";
-import recoverCommand from "./recover.js";
-import repoExplainCommand from "./repo-explain.js";
-import runCommand from "./run.js";
-import startCommand from "./start.js";
-import statusCommand from "./status.js";
-import stopCommand from "./stop.js";
-import { configFilePath } from "../config.js";
-import {
-  initRepoRuntime,
-  MissingWorkflowFileError,
-  parseRepoRuntimeFlags,
-} from "../repo-runtime.js";
-import { resolveRepoRuntimeRoot } from "../orchestrator-runtime.js";
-import { rejectRemovedProjectId } from "../removed-project-id.js";
-import { writeCliError } from "../cli-error.js";
-
-const DOCKER_DEFAULT_CONFIG_DIR = "/var/lib/gh-symphony";
-
-const handler = async (
-  args: string[],
-  options: GlobalOptions
-): Promise<void> => {
-  const [subcommand, ...rest] = args;
-
-  switch (subcommand) {
-    case "init":
-      await repoInit(rest, options);
-      break;
-    case "start":
-      if (rejectRemovedProjectId(rest)) return;
-      await startCommand(rest, repoOptions(options));
-      break;
-    case "run":
-      if (rejectRemovedProjectId(rest)) return;
-      await runCommand(rest, repoOptions(options));
-      break;
-    case "recover":
-      if (rejectRemovedProjectId(rest)) return;
-      await recoverCommand(rest, repoOptions(options));
-      break;
-    case "logs":
-      if (rejectRemovedProjectId(rest)) return;
-      await logsCommand(rest, repoOptions(options));
-      break;
-    case "explain":
-      if (rejectRemovedProjectId(rest)) return;
-      await repoExplainCommand(rest, repoOptions(options));
-      break;
-    case "status":
-      if (rejectRemovedProjectId(rest)) return;
-      await statusCommand(rest, repoOptions(options));
-      break;
-    case "stop":
-      if (rejectRemovedProjectId(rest)) return;
-      await stopCommand(rest, repoOptions(options));
-      break;
-    default:
-      process.stderr.write(
-        "Usage: gh-symphony repo <init|start|status|stop|run|recover|logs|explain> [repo]\n"
-      );
-      process.exitCode = 2;
-  }
-};
-
-export default handler;
-
-export function repoOptions(options: GlobalOptions): GlobalOptions {
-  const repoRuntimeRoot = resolveRepoRuntimeRoot();
-  return {
-    ...options,
-    configDir: shouldUseConfigDirOverride(options, repoRuntimeRoot)
-      ? options.configDir
-      : repoRuntimeRoot,
-  };
-}
-
-function shouldUseConfigDirOverride(
-  options: GlobalOptions,
-  repoRuntimeRoot: string
-): boolean {
-  if (!options.configDirOverride) {
-    return false;
-  }
-
-  if (
-    options.configDirSource === "env" &&
-    options.configDir === DOCKER_DEFAULT_CONFIG_DIR &&
-    existsSync(configFilePath(repoRuntimeRoot))
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
-async function repoInit(args: string[], options: GlobalOptions): Promise<void> {
-  if (rejectRemovedProjectId(args)) {
-    return;
-  }
-
-  let flags: ReturnType<typeof parseRepoRuntimeFlags>;
-  try {
-    flags = parseRepoRuntimeFlags(args);
-  } catch (error) {
-    writeCliError({
-      code: "invalid_arguments",
-      message: error instanceof Error ? error.message : "Invalid arguments",
-      usage:
-        "Usage: gh-symphony repo init [--repo-dir <path>] [--workflow-file <path>]",
-      json: options.json,
-      exitCode: 2,
-    });
-    return;
-  }
-
-  try {
-    const result = await initRepoRuntime({
-      ...flags,
-      ...(options.configDirOverride ? { configDir: options.configDir } : {}),
-    });
-    if (options.json) {
-      process.stdout.write(JSON.stringify(result, null, 2) + "\n");
-      return;
-    }
-    process.stdout.write(
-      [
-        `Repository initialized: ${formatRepoSpec(result.repository)}`,
-        `Runtime: ${result.configDir}`,
-        ...(result.registryProjectId
-          ? [
-              `Config registry: ${options.configDir} (active project: ${result.registryProjectId})`,
-            ]
-          : []),
-        `Workflow: ${result.workflowPath}`,
-      ].join("\n") + "\n"
-    );
-  } catch (error) {
-    writeCliError({
-      code:
-        error instanceof MissingWorkflowFileError
-          ? "missing_workflow_file"
-          : "repository_initialization_failed",
-      message:
-        error instanceof Error
-          ? error.message
-          : "Repository initialization failed.",
-      json: options.json,
-    });
-  }
-}
-
-function formatRepoSpec(repo: { owner: string; name: string }): string {
-  return `${repo.owner}/${repo.name}`;
-}
