+++ packages/orchestrator/src/dispatch.test.ts
@@ -4,10 +4,7 @@ import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { OrchestratorService, sortCandidatesForDispatch } from "./service.js";
-import {
-  explainIssueDispatch,
-  isIssueCandidateEligibleWithReason,
-} from "./explain.js";
+import { isIssueCandidateEligibleWithReason } from "./dispatch-eligibility.js";
 import type {
   OrchestratorTrackerAdapter,
   OrchestratorRunRecord,
@@ -279,404 +276,6 @@ describe("sortCandidatesForDispatch", () => {
 
   it("returns empty array for empty input", () => {
     expect(sortCandidatesForDispatch([])).toEqual([]);
-  });
-});
-
-describe("explainIssueDispatch", () => {
-  const lifecycle = {
-    stateFieldName: "Status",
-    activeStates: ["Todo", "In Progress"],
-    terminalStates: ["Done"],
-    blockerCheckStates: ["Todo"],
-    planningStates: ["Todo"],
-  };
-  const projectRepository = {
-    owner: "acme",
-    name: "repo",
-    cloneUrl: "https://github.com/acme/repo.git",
-  };
-
-  it("explains inactive workflow states", () => {
-    const issue = makeIssue({ identifier: "acme/repo#1", state: "Backlog" });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          id: "workflow_state",
-          status: "block",
-        }),
-      ])
-    );
-  });
-
-  it("explains tracker dispatchability reasons", () => {
-    const issue = makeIssue({
-      identifier: "acme/repo#1",
-      dispatchable: false,
-      dispatchReason: "assigned to another agent",
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.summary).toBe("Not dispatchable: assigned to another agent");
-    expect(report.checks).toContainEqual(
-      expect.objectContaining({
-        id: "tracker_dispatchability",
-        status: "block",
-        message: "not dispatchable: assigned to another agent",
-      })
-    );
-
-    expect(isIssueCandidateEligibleWithReason(issue, lifecycle)).toEqual({
-      eligible: false,
-      reason: "not_dispatchable",
-    });
-    expect(
-      report.checks.find((check) => check.id === "workflow_routability")
-    ).toMatchObject({ status: "pass" });
-  });
-
-  it("explains missing required labels after normalization", () => {
-    const issue = makeIssue({
-      identifier: "acme/repo#1",
-      labels: [" Ready "],
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle: { ...lifecycle, requiredLabels: ["ready", "agent"] },
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toContainEqual(
-      expect.objectContaining({
-        id: "workflow_routability",
-        status: "block",
-        message: 'not routable: Issue is missing required labels ("agent").',
-      })
-    );
-    expect(
-      isIssueCandidateEligibleWithReason(issue, {
-        ...lifecycle,
-        requiredLabels: ["ready", "agent"],
-      })
-    ).toEqual({ eligible: false, reason: "not_routable" });
-  });
-
-  it("prioritizes inactive state over missing required labels", () => {
-    const issue = makeIssue({
-      identifier: "acme/repo#1",
-      state: "Backlog",
-      labels: [],
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle: { ...lifecycle, requiredLabels: ["agent"] },
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.summary).toContain('Project state "Backlog"');
-    expect(report.checks.filter((check) => check.status === "block")[0]?.id).toBe(
-      "workflow_state"
-    );
-  });
-
-  it("explains active linked PR cards when the canonical Issue is inactive", () => {
-    const issue = makeIssue({
-      identifier: "acme/repo#1",
-      state: "In review",
-      metadata: {
-        contentType: "Issue",
-        linkedPullRequests: [
-          {
-            id: "pr-2",
-            number: 2,
-            identifier: "acme/repo#2",
-            url: "https://github.com/acme/repo/pull/2",
-            state: "OPEN",
-            projectState: "In Progress",
-          },
-        ],
-      },
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.summary).toContain("Linked PR card");
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          id: "workflow_state",
-          status: "block",
-          message: expect.stringContaining('canonical Issue state "In review"'),
-          hint: expect.stringContaining(
-            "Linked PR card status alone does not trigger dispatch"
-          ),
-        }),
-      ])
-    );
-  });
-
-  it("explains adapter-derived blocker dispatchability", () => {
-    const issue = makeIssue({
-      identifier: "acme/repo#1",
-      dispatchable: false,
-      dispatchReason: "Blocked by unresolved GitHub issue: acme/repo#2.",
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          id: "tracker_dispatchability",
-          status: "block",
-        }),
-      ])
-    );
-  });
-
-  it("explains already running ownership", () => {
-    const issue = makeIssue({ id: "issue-1", identifier: "acme/repo#1" });
-    const run = makeRun({ runId: "run-1", issueId: issue.id });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [run],
-      activeRunCount: 1,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "runtime_ownership", status: "block" }),
-      ])
-    );
-  });
-
-  it("uses the effective convergence TTL when explaining ownership", () => {
-    const issue = makeIssue({
-      id: "issue-1",
-      identifier: "acme/repo#1",
-      updatedAt: null,
-    });
-    const run = makeRun({
-      runId: "run-convergence",
-      issueId: issue.id,
-      status: "failed",
-      completedAt: "2026-03-09T00:00:00.000Z",
-      runtimeSession: {
-        sessionId: null,
-        threadId: null,
-        status: "completed",
-        startedAt: "2026-03-09T00:00:00.000Z",
-        updatedAt: "2026-03-09T00:00:00.000Z",
-        exitClassification: "convergence-detected",
-      },
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [run],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-      convergenceLock: {
-        now: new Date("2026-03-09T00:02:00.000Z"),
-        ttlMs: 60_000,
-      },
-    });
-
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "runtime_ownership", status: "pass" }),
-      ])
-    );
-  });
-
-  it("does not warn about dirty-workspace recovery for a removed issue workspace", () => {
-    const issue = makeIssue({ id: "issue-1", identifier: "acme/repo#1" });
-    const run = makeRun({
-      runId: "run-incomplete",
-      issueId: issue.id,
-      status: "suppressed",
-      updatedAt: "2026-03-09T00:05:00.000Z",
-      completedAt: "2026-03-09T00:05:00.000Z",
-      issueWorkspaceKey: "workspace-1",
-      recovery: {
-        kind: "incomplete-turn-dirty-workspace",
-        runId: "run-incomplete",
-        issueId: issue.id,
-        issueIdentifier: issue.identifier,
-        workspacePath: "/tmp/work",
-        dirtyFiles: ["partial.txt"],
-        lastEvent: "heartbeat",
-        lastEventAt: "2026-03-09T00:04:30.000Z",
-        sessionId: "session-1",
-        threadId: "thread-1",
-        suggestedCommand: "cd /tmp/work && git status --short && git diff",
-        detectedAt: "2026-03-09T00:05:00.000Z",
-      },
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      issueWorkspaces: [
-        makeIssueWorkspace({
-          workspaceKey: "workspace-1",
-          repositoryPath: "/tmp/work",
-          status: "removed",
-        }),
-      ],
-      runs: [run],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "runtime_ownership", status: "pass" }),
-      ])
-    );
-    expect(JSON.stringify(report)).not.toContain(
-      "dispatch will start a recovery turn"
-    );
-  });
-
-  it("explains project concurrency limits", () => {
-    const issue = makeIssue({ id: "issue-1", identifier: "acme/repo#1" });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 2,
-      maxConcurrentAgents: 2,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "dispatch_limits", status: "block" }),
-      ])
-    );
-  });
-
-  it("explains normalized per-state concurrency limits", () => {
-    const issue = makeIssue({ id: "issue-1", identifier: "acme/repo#1" });
-    const activeRun = makeRun({
-      runId: "run-1",
-      issueId: "issue-2",
-      issueState: " todo ",
-    });
-    const report = explainIssueDispatch({
-      identifier: issue.identifier,
-      issue,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [activeRun],
-      activeRunCount: 1,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: { todo: 1 },
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.checks).toContainEqual(
-      expect.objectContaining({
-        id: "dispatch_limits",
-        status: "block",
-        details: expect.objectContaining({ activeInState: 1, stateLimit: 1 }),
-      })
-    );
-  });
-
-  it("prioritizes repository linkage in the not-found summary", () => {
-    const report = explainIssueDispatch({
-      identifier: "other/repo#1",
-      issue: null,
-      projectRepository,
-      lifecycle,
-      issueRecords: [],
-      runs: [],
-      activeRunCount: 0,
-      maxConcurrentAgents: 3,
-      maxConcurrentAgentsByState: {},
-    });
-
-    expect(report.dispatchable).toBe(false);
-    expect(report.summary).toContain(
-      "Repository other/repo is not the active managed project repository"
-    );
   });
 });
