+++ packages/orchestrator/src/index.ts
@@ -24,7 +24,7 @@ export { OrchestratorService, createStore };
 export { resolveAdaptivePollIntervalMs } from "./service.js";
 export type { OrchestratorLogLevel };
 export * from "./runtime-factory.js";
-export * from "./explain.js";
+export * from "./dispatch-eligibility.js";
 export * from "./repository-cache.js";
 export {
   findGithubProjectIssue,
